requirements.txt
google-generativeai
srt